+++ lib/main.dart
@@ -19,7 +19,6 @@ import 'package:da_kanji_mobile/CodegenLoader.dart';
 import 'package:da_kanji_mobile/dakanji_app.dart';
 import 'package:da_kanji_mobile/feedback_localization.dart';
 import 'package:da_kanji_mobile/globals.dart';
-import 'package:da_kanji_mobile/init.dart';
 import 'package:da_kanji_mobile/widgets/widgets/dakanji_splash.dart';
 
 Future<void> main() async {